fixed middleware file location to work correctly, and ensured JWT functionality is in place. Dashboard routes are now protected with authentication.
