+++ src/lib/auth.config.ts
@@ -1,6 +1,6 @@
 import type { NextAuthConfig } from 'next-auth';
  
-export const authConfig = {
+export const authConfig: NextAuthConfig = {
   pages: {
     signIn: '/login',
   },
